QOL features added QuickClose when exiting from homepage form if QuickClose is enabled the application will close from there removing the need to close from the login page
NEW FEATURE
the homepage is almost here 1/6 of it has been complete huzzar
in this update the todo column makes an appearance allowing you to quickly edit your todo notes!
you can also move them with the press of a button or delete them if they're no longer needed
this adds huge useability to the application and allows for the ability to quickly manage the tasks you still have to do

What you can look forward to!
the next major update will be the coming up tab on the home page (top right) this'll allow you to view your tasks that are comeing up and need to be completed, this will allow you to better priorities workloads and better understand what needs to be done

in the near future more boards will be comeing! yes you heard me right soon you'll have more than one board which is fantastic for you, this feature may take a while to come out because of how the application was made it only cares about the one board so all that will have to change, this also means the entire code base will be getting revamped and cleaned up removing any code that is not needed

and finally custom colour themes, these have been in the works and background for a while now with the layout being designed for them, most of the hard work is done i've just got to finish it up and then you'll be all set